pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agidigraph"
version = "0.1.0"
description = "LLM-driven climate disaster impact graph builder and explorer"
requires-python = ">=3.7"
authors = [{ name = "Hackathon Team" }]
readme = "README.md"
dependencies = [
    "fastapi>=0.103",
    "uvicorn>=0.22",
    "langgraph>=0.2.28",
    "langchain-openai>=0.1.7",
    "pandas>=1.3",
    "networkx>=2.5",
    "pydantic>=2.5",
    "pydantic-settings>=2.0",
    "python-dotenv>=0.19",
    "streamlit>=1.20",
    "pyvis>=0.3",
    "httpx>=0.24",
    "openai>=1.3.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.23",
    "ruff>=0.5"
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "ANN"]

[tool.setuptools.packages.find]
include = ["app*", "scripts*", "tests*", "ui*"]

[tool.pytest.ini_options]
addopts = "-ra"
python_files = "tests/*.py"
